Changed Property Names and Types
Customer: Changed "CustomerID" to "CustomerId" to be consistent with the Id fields of the other classes.

Order: Changed "orderDate" to "OrderDate", a necessary change as Properties must start with a Capital letter.

Order Item: Changed the type of the PurchasePrice property from "decimal?" to "Decimal?" so that it is consistent with the other Decimal types in other classes.