Add new classes to separate responsibilities
The biggest change is that all of the persistence methods (retrieve and save) were removed from the Customer, Order, and Product classes and moved into repository classes.

An Address class was created to delegate the responsibility of handling the many properties and formatting of addresses. This will be used by the Customer and Order classes.