Showcased creating a reusable component
Created the Acme.Common component to showcase the creation of a component that could be used by other projects.

Within that component I created a StringHandler class. This class was at first created as an instance class with a public InsertSpaces method.

To showcase static methods, the class was made static, as was the InsertSpaces method.

Taking this one step further, I made the InsertSpaces method an extension method by adding the "this" keyword in the parameter list for the source string. This allows me to use the InsertSpaces method directly as a method of the string class and see it within the IntelliSense window for any string variables I create.